Split artifact Playwright flows
