+++ E2E/playwright/tests/artifacts.spec.ts
@@ -1,0 +1,146 @@
+import { test, expect } from '@playwright/test';
+import { clickSidebarTool, harnessURL } from './harness-helpers';
+
+test('mock harness surfaces file artifacts from tool cards', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await page.getByLabel('Message').fill('Can you write a file that says hello world');
+  await page.getByRole('button', { name: 'Send' }).click();
+
+  await expect(page.getByTestId('tool-card-title')).toHaveText('host.file.write');
+  await expect(page.getByTestId('tool-card-artifacts')).toBeVisible();
+  await expect(page.getByTestId('tool-card-artifact-label')).toHaveText('hello.txt');
+  await expect(page.getByTestId('tool-card-artifact-detail')).toHaveText('/mock/QuillCode');
+  await expect(page.getByTestId('tool-card-artifact')).toHaveAttribute('data-kind', 'file');
+  await expect(page.getByTestId('tool-card-artifact')).toHaveAttribute('href', 'file:///mock/QuillCode/hello.txt');
+  await expect(page.getByTestId('tool-card-text-previews')).toBeVisible();
+  await expect(page.getByTestId('tool-card-text-preview-label')).toHaveText('hello.txt');
+  await expect(page.getByTestId('tool-card-text-preview-content')).toHaveText('hello world');
+  await expect.poll(() => page.getByTestId('tool-card-details').evaluate(element => (element as HTMLDetailsElement).open)).toBe(false);
+  await page.getByTestId('tool-card-details').locator('summary').click();
+  await expect.poll(() => page.getByTestId('tool-card-details').evaluate(element => (element as HTMLDetailsElement).open)).toBe(true);
+  await expect(page.getByTestId('tool-card-output')).toContainText('/mock/QuillCode/hello.txt');
+  await expect(page.getByText('Wrote `hello.txt`.')).toBeVisible();
+
+  await clickSidebarTool(page, 'activity-button');
+  await expect(page.getByTestId('activity-pane')).toBeVisible();
+  await expect(page.getByTestId('activity-task-title')).toContainText('Can you write a file');
+  await expect(page.getByTestId('activity-tool')).toContainText('host.file.write');
+  await expect(page.getByTestId('activity-artifact')).toContainText('hello.txt');
+  await expect(page.getByTestId('activity-artifact')).toContainText('/mock/QuillCode');
+  await expect(page.getByTestId('activity-artifact')).not.toContainText('undefined');
+  await expect(page.getByTestId('activity-source').first()).toContainText('AGENTS.md');
+  await expect(page.getByTestId('activity-plan')).toHaveCount(5);
+  await expect(page.getByTestId('activity-plan').nth(0)).toContainText('Understand request');
+  await expect(page.getByTestId('activity-plan').nth(2)).toContainText('Use tools');
+  await expect(page.getByTestId('activity-plan').nth(2)).toContainText('Done');
+  await expect(page.getByTestId('activity-plan').nth(4)).toContainText('Answer user');
+  await expect(page.getByTestId('activity-handoff')).toContainText('Thread: Can you write a file');
+  await expect(page.getByTestId('activity-handoff')).toContainText('Tools: 1 tool (host.file.write)');
+  await expect(page.getByTestId('activity-handoff')).toContainText('Artifacts: 1 artifact (hello.txt)');
+  await expect(page.getByTestId('activity-handoff')).not.toContainText('\\n');
+  await expect(page.getByTestId('activity-final-answer')).toContainText('Wrote `hello.txt`.');
+
+  await page.getByTestId('activity-handoff-section').getByTestId('activity-section-toggle').click();
+  await expect(page.getByTestId('activity-handoff-section')).toHaveAttribute('data-collapsed', 'true');
+  await expect(page.getByTestId('activity-handoff')).toHaveCount(0);
+  await page.getByTestId('activity-handoff-section').getByTestId('activity-section-toggle').click();
+  await expect(page.getByTestId('activity-handoff')).toContainText('Latest answer: Wrote `hello.txt`.');
+
+  await page.getByTestId('activity-plan-section').getByTestId('activity-section-toggle').click();
+  await expect(page.getByTestId('activity-plan-section')).toHaveAttribute('data-collapsed', 'true');
+  await expect(page.getByTestId('activity-plan')).toHaveCount(0);
+  await page.getByTestId('activity-plan-section').getByTestId('activity-section-toggle').click();
+  await expect(page.getByTestId('activity-plan')).toHaveCount(5);
+
+  await page.getByTestId('activity-tool-section').getByTestId('activity-section-toggle').click();
+  await expect(page.getByTestId('activity-tool-section')).toHaveAttribute('data-collapsed', 'true');
+  await expect(page.getByTestId('activity-tool')).toHaveCount(0);
+  await page.getByTestId('activity-tool-section').getByTestId('activity-section-toggle').click();
+  await expect(page.getByTestId('activity-tool-section')).toHaveAttribute('data-collapsed', 'false');
+  await expect(page.getByTestId('activity-tool')).toContainText('host.file.write');
+});
+
+test('mock harness renders image artifact previews from tool cards', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await page.getByLabel('Message').fill('take a screenshot');
+  await page.getByRole('button', { name: 'Send' }).click();
+
+  await expect(page.getByTestId('tool-card-title')).toHaveText('host.computer.screenshot');
+  await expect(page.getByTestId('tool-card-artifact-label')).toHaveText('screenshot.png');
+  await expect(page.getByTestId('tool-card-image-previews')).toBeVisible();
+  await expect(page.getByTestId('tool-card-image-preview')).toBeVisible();
+  await expect(page.getByTestId('tool-card-image-preview')).toHaveAttribute('data-kind', 'image');
+  await expect(page.getByTestId('tool-card-image-preview-type')).toHaveText('Image · PNG');
+  await expect(page.getByTestId('tool-card-image-preview-label')).toHaveText('screenshot.png');
+  await expect(page.getByTestId('tool-card-image-preview-detail')).toHaveText('/mock/QuillCode/screenshots');
+  await expect(page.getByTestId('tool-card-image-preview').locator('img')).toHaveAttribute('src', 'file:///mock/QuillCode/screenshots/screenshot.png');
+  const imageSurface = await page.getByTestId('tool-card-image-preview').evaluate((element) => {
+    const cardStyle = getComputedStyle(element);
+    const imageStyle = getComputedStyle(element.querySelector('img')!);
+    return {
+      cardRadius: cardStyle.borderRadius,
+      imageRadius: imageStyle.borderRadius,
+      imageOutlineColor: imageStyle.outlineColor,
+      imageOutlineWidth: imageStyle.outlineWidth,
+      imageOutlineOffset: imageStyle.outlineOffset
+    };
+  });
+  expect(imageSurface.cardRadius).toBe('18px');
+  expect(imageSurface.imageRadius).toBe('10px');
+  expect(imageSurface.imageOutlineColor).toBe('rgba(255, 255, 255, 0.1)');
+  expect(imageSurface.imageOutlineWidth).toBe('1px');
+  expect(imageSurface.imageOutlineOffset).toBe('-1px');
+  await expect(page.getByText('Captured a screenshot (1280 x 720).')).toBeVisible();
+});
+
+test('mock harness renders document artifact previews from tool cards', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await page.getByLabel('Message').fill('make a pdf artifact');
+  await page.getByRole('button', { name: 'Send' }).click();
+
+  await expect(page.getByTestId('tool-card-title')).toHaveText('host.file.write');
+  await expect(page.getByTestId('tool-card-artifact-label')).toHaveText('briefing.pdf');
+  await expect(page.getByTestId('tool-card-document-previews')).toBeVisible();
+  await expect(page.getByTestId('tool-card-document-preview')).toBeVisible();
+  await expect(page.getByTestId('tool-card-document-preview')).toHaveAttribute('data-kind', 'pdf');
+  await expect(page.getByTestId('tool-card-document-preview-type')).toHaveText('PDF · PDF');
+  await expect(page.getByTestId('tool-card-document-preview-label')).toHaveText('briefing.pdf');
+  await expect(page.getByTestId('tool-card-document-preview-detail')).toHaveText('/mock/QuillCode/reports');
+  await expect(page.getByTestId('tool-card-document-preview-open')).toHaveAttribute('href', 'file:///mock/QuillCode/reports/briefing.pdf');
+  const documentSurface = await page.getByTestId('tool-card-document-preview').evaluate((element) => {
+    const cardStyle = getComputedStyle(element);
+    const iconStyle = getComputedStyle(element.querySelector('.artifact-document-icon')!);
+    return {
+      cardRadius: cardStyle.borderRadius,
+      cardMinHeight: cardStyle.minHeight,
+      iconRadius: iconStyle.borderRadius,
+      transitionProperty: cardStyle.transitionProperty
+    };
+  });
+  expect(documentSurface.cardRadius).toBe('18px');
+  expect(documentSurface.cardMinHeight).toBe('74px');
+  expect(documentSurface.iconRadius).toBe('10px');
+  expect(documentSurface.transitionProperty).toBe('transform, box-shadow');
+  await expect(page.getByText('Created `briefing.pdf`.')).toBeVisible();
+});
+
+test('mock harness renders appshot artifact previews from tool cards', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await page.getByLabel('Message').fill('make an appshot artifact');
+  await page.getByRole('button', { name: 'Send' }).click();
+
+  await expect(page.getByTestId('tool-card-title')).toHaveText('host.appshot.capture');
+  await expect(page.getByTestId('tool-card-artifact-label')).toHaveText('checkout.appshot.json');
+  await expect(page.getByTestId('tool-card-document-previews')).toBeVisible();
+  await expect(page.getByTestId('tool-card-document-preview')).toHaveAttribute('data-kind', 'appshot');
+  await expect(page.getByTestId('tool-card-document-preview-type')).toHaveText('Appshot · APPSHOT');
+  await expect(page.getByTestId('tool-card-document-preview-label')).toHaveText('checkout.appshot.json');
+  await expect(page.getByTestId('tool-card-document-preview-detail')).toHaveText('/mock/QuillCode/appshots');
+  await expect(page.getByTestId('tool-card-document-preview-open')).toHaveAttribute('href', 'file:///mock/QuillCode/appshots/checkout.appshot.json');
+  await expect(page.getByTestId('tool-card-text-previews')).toHaveCount(0);
+  await expect(page.getByText('Captured appshot `checkout.appshot.json`.')).toBeVisible();
+});
